requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
python-dotenv
google-generativeai
pydantic
